MA-3945-Improve height precision and rounding for BLE profile synchronization and BMI calculations.
- **ConversionTools**: Update `convertStoredHeightToCm` to return a `Double` instead of an truncated `Int` to preserve precision during unit conversion.
- **Account / SettingsViewModel**: Implement manual rounding of centimeter values before converting to `GGBTUserProfile`. This prevents the BT SDK's internal `toInt()` truncation from causing off-by-one errors (e.g., ensuring 182.88cm rounds to 183cm rather than truncating to 182cm).
- **EntryReducer / EntryHelper**: Apply `round()` to height values before passing them to BMI calculation functions to maintain consistency with the updated storage and transmission logic.

+++ Android/app/src/main/java/com/dmdbrands/gurus/weight/core/shared/utilities/ConversionTools.kt
@@ -93,8 +93,8 @@ object ConversionTools {
    * @param stored Height in stored format (tenths of inches)
    * @return Height in centimeters
    */
-  fun convertStoredHeightToCm(stored: Int): Int {
-    return (stored * CM_TO_INCH_FACTOR).toInt()
+  fun convertStoredHeightToCm(stored: Int): Double {
+    return (stored * CM_TO_INCH_FACTOR)
   }
 
   /**
